[CL-2308] Show input label cropping image. Make exports consistent
